+++ db/upgrade.php
@@ -167,6 +167,23 @@ function xmldb_format_grid_upgrade($oldversion = 0) {
         upgrade_plugin_savepoint(true, 2022072200, 'format', 'grid');
     }
 
+    if ($oldversion < 2023051001) {
+        $records = $DB->get_records('course_format_options',
+            array(
+                'format' => 'grid',
+                'name' => 'numsections'
+            ), '', 'id'
+        );
+
+        $records = array_keys($records);
+        foreach ($records as $id) {
+            $DB->set_field('course_format_options', 'name', 'gnumsections', array('id' => $id));
+        }
+
+        // Grid savepoint reached.
+        upgrade_plugin_savepoint(true, 2023051001, 'format', 'grid');
+    }
+
     // Automatic 'Purge all caches'....
     purge_all_caches();