+++ Changes.md
@@ -4,6 +4,7 @@ Version 402.1.1 - TBR
 ----------------------------
 1. Fix 'Previous/Next section arrows are incorrect for RTL languages' - #183.
 2. Fix 'Last section not showing its image when editing'.
+3. Convert 'numsections' to 'gnumsections' to facilitate both a course setting and flexible section management when editing.
 
 Version 402.1.0 - 11/05/2023
 ----------------------------